requirements.txt
fastapi
uvicorn
razorpay
python-dotenv
supabase
resend
